pyproject.toml
[project]
name = "wind-ml-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "dvc>=3.30.0",
    "joblib>=1.3.0",
    "matplotlib>=3.7.0",
    "mlflow>=2.12.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "xgboost>=2.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]
extend-ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.3.0",
    "pytest>=8.0.0",
]
